% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19RR-package.R
\docType{package}
\name{Covid19RR-package}
\alias{Covid19RR}
\alias{Covid19RR-package}
\title{Covid19RR: Estimation of Reproduction Rates for the COVID-19 Pandemic in Denmark based on Publicly Available Data}
\description{
What the package does (one paragraph).
}
\examples{
dat <- preprocess.data(download.data())
obj <- setup.TMB.object(dat)
opt <- fit(obj,fix=c(logrsigma=-6))
plot_fit(dat,opt)
save.image()

# Or to launch the shiny whatsit:
\dontrun{
launch_shiny("RRest")
}
}
\author{
\strong{Maintainer}: Uffe Thygesen \email{uhth@dtu.dk}

Authors:
\itemize{
  \item Matt Denwood \email{md@sund.ku.dk}
}

}
\keyword{internal}
