% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv19rr.R
\name{cv19rr}
\alias{cv19rr}
\alias{print.cv19rr}
\alias{plot.cv19rr}
\alias{autoplot.cv19rr}
\title{S3 methods for the cv19rr class}
\usage{
\method{print}{cv19rr}(x, ...)

\method{plot}{cv19rr}(
  x,
  gen_time = 4.7,
  tests = 50000,
  lag = 7,
  ylim = c(0.5, 1.5),
  page = 3,
  pngfile = NULL,
  ...
)

\method{autoplot}{cv19rr}(object, ...)
}
\arguments{
\item{x}{An object of class cv19rr}

\item{...}{Other arguments provided for compatibility with S3 methods}

\item{gen_time}{Generation time (growth?) parameter}

\item{tests}{Number of tests to normalise against}

\item{lag}{Lag parameter}

\item{ylim}{Limits for the y axis of the Kontakttal R plot}

\item{page}{Which page of plots to display (1, 2, 3 or multiple)}

\item{pngfile}{An optional pngfile to use}

\item{object}{An object of class cv19rr}
}
\description{
S3 methods for the cv19rr class
}
