% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Title}
\usage{
download_data(preprocess = TRUE)
}
\arguments{
\item{preprocess}{Should the downloaded data be processed automatically?}
}
\value{
dataset

#TODO clean up tempfile
}
\description{
Title
}
