% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cv19rr.R
\name{estimate_cv19rr}
\alias{estimate_cv19rr}
\title{Estimation of RR}
\usage{
estimate_cv19rr(dat, parameters = NULL, fix = NULL, ...)
}
\arguments{
\item{dat}{Incidence data}

\item{parameters}{Option to override the default parameter values}

\item{fix}{Option to override the default fixed values}

\item{...}{Other arguments provided for compatibility with \code{\link[EpiEstim]{estimate_R}} but currently ignored}
}
\value{
An object of class \link{cv19rr}
}
\description{
Estimation of RR
}
\details{
Intended to have a similar interface to \code{\link[EpiEstim]{estimate_R}}
}
\examples{
dat <- download_data()

}
