% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Launch an interactive Shiny app contained within this package}
\usage{
launch_shiny(appname, ...)
}
\arguments{
\item{appname}{the name of an app to launch (see the details section)}

\item{...}{further arguments to be passed to \code{\link[shiny]{runApp}}}
}
\description{
This function provides a convinient way in which to launch one of the interactive shiny apps contained within this package
}
\details{
TODO
The apps that are currently available are:
}
\examples{
\dontrun{
launch_shiny('RRest')
}

}
