% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19RR-package.R
\docType{package}
\name{Covid19RR-package}
\alias{Covid19RR}
\alias{Covid19RR-package}
\title{Covid19RR: Estimation of Reproduction Rates for the COVID-19 Pandemic in Denmark based on Publicly Available Data}
\description{
Implementation of a model used by the expert modelling group for use
            by SSI and others.
}
\examples{
## Using the interface:
dat <- download_data()
mod <- estimate_cv19rr(dat)
# Or with non-default parameters or fix:
mod <- estimate_cv19rr(dat, parameters=list(beta=1), fix=c(logrsigma=-6))
plot(mod)
save.image()

## Old version using internal (now non-exported) code:
obj <- Covid19RR:::setup.TMB.object(dat)
opt <- Covid19RR:::fit(obj,fix=c(logrsigma=-6))
Covid19RR:::plot.cv19rr(list(dat=dat,opt=opt))
save.image()

# Or to launch the shiny whatsit:
\dontrun{
launch_shiny("RRest")
}

# Or to read the vignette:
\dontrun{
vignette("Covid19RR")
}
}
\author{
\strong{Maintainer}: Uffe Thygesen \email{uhth@dtu.dk}

Authors:
\itemize{
  \item Matt Denwood \email{md@sund.ku.dk}
}

}
\keyword{internal}
