% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19RR-package.R
\docType{package}
\name{Covid19RR-package}
\alias{Covid19RR}
\alias{Covid19RR-package}
\title{Covid19RR: Estimation of Reproduction Rates for the COVID-19 Pandemic in Denmark based on Publicly Available Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Uffe Thygesen \email{uhth@dtu.dk}

Authors:
\itemize{
  \item Matt Denwood \email{md@sund.ku.dk}
}

}
\keyword{internal}
