% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19RR-package.R
\docType{package}
\name{Covid19RR-package}
\alias{Covid19RR}
\alias{Covid19RR-package}
\title{Covid19RR: Estimation of Reproduction Rates for the COVID-19 Pandemic in Denmark based on Publicly Available Data}
\description{
Implementation of a model used by the expert modelling group for use
            by SSI and others.
}
\examples{
## Using the interface:
dat <- download_data()
mod <- estimate_cv19rr(dat)
plot(mod)
# Or with non-default parameters or fix:
mod <- estimate_cv19rr(dat, parameters=list(beta=0.7))
plot(mod)
mod <- estimate_cv19rr(dat, fix=c(beta=0.7))
plot(mod)

# Or to launch the shiny whatsit:
\dontrun{
launch_shiny("RRest")
}

# Or to read the vignette:
\dontrun{
vignette("Covid19RR")
}
}
\author{
\strong{Maintainer}: Uffe Thygesen \email{uhth@dtu.dk}

Authors:
\itemize{
  \item Matt Denwood \email{md@sund.ku.dk}
}

}
\keyword{internal}
