% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv19rr.R
\name{cv19rr}
\alias{cv19rr}
\alias{print.cv19rr}
\alias{plot.cv19rr}
\alias{autoplot.cv19rr}
\title{S3 methods for the cv19rr class}
\usage{
\method{print}{cv19rr}(x, ...)

\method{plot}{cv19rr}(x, pngfile = NULL, ...)

\method{autoplot}{cv19rr}(object, ...)
}
\arguments{
\item{x}{An object of class cv19rr}

\item{...}{Other arguments provided for compatibility with S3 methods}

\item{pngfile}{An optional pngfile to use}

\item{object}{An object of class cv19rr}
}
\description{
S3 methods for the cv19rr class
}
